Rename setLocalStorage file and adjust paths
